subnational_carbon_budget_assessments/requirements.txt
django
pandas
djangorestframework
markdown